organize project structure a bit
