+++ package.json
@@ -2,7 +2,7 @@
   "name": "ipfs-merkle-dag",
   "version": "0.0.1",
   "description": "A JavaScript implementations of the IPFS MerkleDAG implementations (protobufs)",
-  "main": "./node.js",
+  "main": "./src/index.js",
   "scripts": {
     "test": "node tests/index.js",
     "lint": "standard"
